Make convert errors more clear to users
Fixes https://github.com/pulumi/pulumi/issues/13093

This makes two changes to how we print diagnostics in convert.
Firstly if program generation returns any warning diagnostics we only
print those if either PULUMI_DEV is true, or if we're output pcl (which
is always dev-focused).
Secondly if we have any error diagnostics we print an error similar to
the panic handler making trying to make it very clear to the user that
the error is a bug. We disable this message if using --strict as in that
case while there may still be codegen errors that really are bugs there
will also be many PCL errors that should be errors when converting with
--strict set.