Python codegen prints input deprecation message iff the input is used.
Python codegen set the input to the default value and then checked if it
was used to determine if the deprecation message should be printed.

This caused the deprecation message to always be printed if a default
value was set.

This fixes this and is split into 2 commits for easier reviewing.