Updates the error handling logic around RPC deserialization in Python
Today, if the resolved output value for a resource does not match exactly
the expected type for that output, deserialization will fail on an unhandled exception.
For sum types where an output can be either a map or a list, we do not fall back,
so the program will not complete. This PR changes the exception raising logic to
simply log an error when parsing. This will allow the program to fall back,
and continue running.