[sdk/nodejs] Write port as string to avoid colorized output
Node's `console.log` colors numbers by default if it thinks that the output is a terminal (unless `NO_COLOR` is set). At least one user is running into a case where the port is being outputted surrounded with color codes, when using a dynamic provider on v20.3.0. To mitigate, we can update the places where we use `console.log(port)`, converting the number to a string first.