Fix diff display partially parsing JSON/YAML
Fixes https://github.com/pulumi/pulumi/issues/10312

When attempting to decode a string property value as JSON or YAML the
decoder would return a partial parse as valid. This resulted in strings
like "{} stuff" being reported as an empty JSON object, rather than the
literal string "{} stuff".

This also fixes the decoder to ignore whitespace around the value so for
example "   [1, 2]   " is now treated as a JSON array, not a string
literal.