[sdk/python] remove deprecation warning on result types.
This change causes deprecation messages on Result objects from Function
invokes to only be shown when using getters on the result object instead
of on instantiation.

SDK functions return Result objects and they may have deprecated fields.
Instantiation of these objects causes deprecation messages to appear
even if a user is not using the field.

Fixes #11739