CLI update summary now shows after non-mandatory policy events seen.
This change causes only mandatory policy violations to skip the summary
display causing the resources changed and duration details to not be
printed.

Policy Violations are categorized as `mandatory`, `disabled`, and
`advisory`. We were treating all 3 as policy violations to skip
displaying the update summary. Now we only use `mandatory`.