Remove the downstream codegen tests
These have been failing for months now and there's no real apatite to
fix them up.

They have always been problematic, and have failed to catch a number of
significant codegen regressions because it was always on the PR author
to go and read the _very large_ diffs and see if it looked ok.

We'll be covering codegen (and runtime) correctness with matrix testing.
There's no need to also couple our testing to every downstream provider,
and if we do want to do that it should be a proper test. Not a manual
diff comparison.