Add a better error message for invalid NodeJS AutoAPI workdir
If you supply a workDir to a NodeJS AutoAPI LocalWorkspace, it
does not validate its correctness, and proceeds to use it when
trying to spawn commands. That leads to misleading and confusing
"Command failed with ENOENT: pulumi version" error messages,
making it look like the issue is a missing `pulumi` binary on
your PATH. This changes to eagerly check and issue a more
friendly error message. Fixes #13274.