WIP: Add a Nix flake
This adds the first pieces of a Nix flake to the Pulumi repository.
We hope to be able to use this to simplify the local devleopment
process.
It's intended to be used almost exclusively for development--once
usable.

The Flake defines a devShell that isn't yet usable,
and a `pulumi-cli` package that builds the binary -- this works.
To use:

```
% nix build '.#pulumi-cli'
% result/bin/pulumi
```

Use `nix shell` to get a shell environment with that binary.

```
% nix shell '.#pulumi-cli'
```

This is nowhere near ready for review or usage yet.