test: Attempt to fix TestCreatingProjectWithEmptyConfig
TestCreatingProjectWithEmptyConfig appears to be broken on master:

```
=== Failed
=== FAIL: cmd/pulumi TestCreatingProjectWithEmptyConfig (0.00s)
    new_test.go:330:
        	Error Trace:	/home/runner/work/pulumi/pulumi/pkg/cmd/pulumi/new_test.go:330
        	Error:      	Received unexpected error:
        	            	A project with this name already exists
        	Test:       	TestCreatingProjectWithEmptyConfig

=== FAIL: cmd/pulumi TestCreatingProjectWithEmptyConfig (re-run 1) (0.28s)
    new_test.go:330:
        	Error Trace:	/home/runner/work/pulumi/pulumi/pkg/cmd/pulumi/new_test.go:330
        	Error:      	Received unexpected error:
        	            	A project with this name already exists
        	Test:       	TestCreatingProjectWithEmptyConfig
```

As seen in #13323 and the attempts to land #13322.

To reduce likelihood of "name already exists" errors,
generate a random name for all tests where the name is currently based
on the name of the temporary directory.

Also, change all manually managed temporary directories
(with an unchecked error) to `t.TempDir()` -- making the testing package
handle cleanup of the temporary directories.