Temp changes to unblock the release
Publishing v3.74.0 to npm failed with `../../scripts/publish_npm.sh: line 13: my_array[2]: unbound variable`.

This commit temporarily changes the `publish_npm.sh` script to not have the unbound variable and comments out jobs in the release workflow that already succeeded, and adds a way to manually dispatch the release workflow to unblock the release. Will use that to re-run the release workflow.

Afterwords, we can revert and address the actual issue.