Fix `publish_npm.sh` script to handle `GITHUB_REF` without '/'
It looks like the intention of this script was to parse `GITHUB_REF`s in the form `refs/heads/<branch_name>` or `refs/tags/<tag_name>` per https://docs.github.com/en/actions/learn-github-actions/variables#default-environment-variables, but we overwrite this envvar in the workflow:

https://github.com/pulumi/pulumi/blob/53e205f08fc08f3388e2a63d0335b8e3a8c93a71/.github/workflows/release.yml#L54

Which is passed along from here:

https://github.com/pulumi/pulumi/blob/53e205f08fc08f3388e2a63d0335b8e3a8c93a71/.github/workflows/on-release.yml#L37

Which is just the tag name.

Still unclear how this worked for previous releases, but the script errored with the latest release when `GITHUB_REF` was just `v3.74.0`.

This commit makes the script resilient to values of `GITHUB_REF` that don't have '/'.