sdk/go/pulumix: Add {Ptr,Array,Map}Output and specialization
Adds generic output containers for pointers, arrays, and maps.
In addition to the base output types, this includes `G` (as in generic)
variants of these that allow specifying the kind of "inner" output.

For example, if we have a `pux.PtrOutput[s3.Bucket]`,
calling `.Elem()` on it will give us back a `pux.Output[s3.Bucket]`.
This will not have the accessor methods that `s3.BucketOutput` includes:

    func (o BucketOutput) Arn() pulumi.StringOutput
    func (o BucketOutput) Bucket() pulumi.StringOutput

Which means that a user will have to rely on `pux.Apply`.
To avoid that, the GPtrOutput allows specifying the kind of value
that will be returned by `Elem`.

The Elem method on `pux.GPtrOutput[s3.Bucket, s3.BucketOutput]` will
return an `s3.BucketOutput` so a user will be able to do:

    o.Elem().Arn()

Generic variants are included for all three container types.

This required a new pux function: SpecializeOutput that
turns a value of one output type into another output type
as long as they have the same element type.