Fix http source URL interpolation
Turns out this wasn't tested and regressed in
https://github.com/pulumi/pulumi/commit/51cc6461df930b17e23d2d1408cc6110b1ede7be
when we started URL parsing the input string and so ended up with a text
string with "%7B" instead of "{" to replace.

This fixes the interpolation and changes a couple of the http tests to
actually use this feature so it's tested.