Add regression test for #13301
Adds a regression of the bug using code provided by @phillipedwards.
The reproduction is a bit complicated because of the very specific
scenario we need to replicate here.

Additionally, we need an untidy go.mod file for this,
so we need to teach scripts/tidy.sh how to ignore files.