[engine] Patch MLC source positions
When a user program creates a multi-lanugage component, the engine
issues a construct call to the component's provider rather than
issuing an actual resource registration. Instead, the component's
provider is responsible for registering the component resource. This
creates a problem for source positions: because the component provider
is registering the component, the component's source location will be
recorded as the position of the registration in the component provider
rather than in the user program.

These changes address this problem by starting a proxy per component
instantiation that patches the source position of the MLC itself when
it is registered and clears the source positions for other resources
registered by the MLC.

Instead of clearing source positions for non-root resources registered
by MLCs, future changes should add support for translating the reported
positions for these resources to component-relative positions.