Add the ability to verify an expected log message was written
We used to test that various errors we're written out to the user in
these langhost tests, but at some point those test were removed (I
suspect because we stopped printing them to stderr).

This adds in the ability to check for a log message in the engine
diagnostics stream, and makes use of that in one of the tests.