Rework InstallDependencies and RunPlugin output
Both of these commands need to send output back to the engine host, but
they did it in slightly different ways. We also need output from the
"Pack" command when we add that, so we'd then have three commands
needing to send output back to the engine.

This unifies the commands to use a new grpc service that acts a bit like
a terminal, and we can wrap it in the language plugins to behave as a
tty.