test(codegen): Disable regress-11176
regress-11176 is currently broken because it imports awsx which,
without https://github.com/pulumi/pulumi-awsx/pull/1066 released
pulls in an old version of google.golang.org/genproto,
and that conflicts with the newer version.

Details in https://github.com/pulumi/pulumi/issues/13644#issuecomment-1664711242

This change disables the test so that tests on master aren't red.

Refs #13644