ProgramTest: Always write output to log file
I'm having trouble debugging some test timeouts
and part of the problem is that the command outputs
are not recorded to a file until after the command finishes running.

To help debug problems such as this,
I've changed RunCommand to always create the output log file,
and to stream output to it as the command runs.
So even if the test times out and we aren't able to print anything,
it'll report where the log file resides.

On top of that, RunCommand retains its original behavior
with some minor QoL improvements:

- In verbose mode, if either stdout or stderr is unset,
  that stream is written to the testing.T log so it's not lost.
- In non-verbose mode, output is always written to the testing.T.
  This has two nice properties:
  if the test fails, logs are correlated to the failing test,
  and if `go test -v` is used, output is streamed to stderr directly.
  These properties of `t.Logf` also obviate the need
  to print the log output to stderr explicitly if the command failed.
- RunCommand takes a testing.TB instead of a testing.T.
  This allows us to test these beahviors.

Note that to retain the old expectations,
RunCommand still has to write to stdout/stderr in non-verbose mode
if the command failed.