Lock access to the plugin loading channels
Fixes https://github.com/pulumi/pulumi/issues/12371.

This locks access to the plugin request channels with a RWLock.
Before trying to write to the channel we try to take a Read lock (yes
this sounds the wrong way round, carry on). Many loaders are free to
send to the loadRequest channel at once, but we use the read lock to
atomiclly track if any are currently in progress.

When we go to close the plugin host the first thing we do is take a
Write lock. Firstly this can't be taken until all the read locks are
released indicating that no plugins are currently loading, but secondly
while the write lock is taken no more read locks can be taken blocking
any further plugin loads from starting.

We never release this write lock, thus permenatly blocking plugin loads
once `Close` is called. So that we don't indefinently block inside load
calls we do a `TryRLock` and return an error if the read lock can't be
taken.

With this locking in place the rest of `Close` is then free to shut down
all current plugins and close of the request channels, assured that they
shouldn't be posted to again as the lock stays held.