ci: Don't upload coverage artifacts on PRs
We currently upload coverage artifacts to codecov every 12 hours via the periodic-coverage cron workflow, which runs the full suite of tests against master, as well as during PRs, which runs a subset of tests for the PR.

We currently see some discrepancies in the data in codecov. One theory is the difference in tests between PRs and master.

This change makes it so we no longer upload coverage artifacts during PRs, to helpfully keep the data in codecov clean for trends over time.