feat: Allow testify-style assertions
Currently, assertions are hand written as functions
that return `(ok bool, msg string)`.
Long-term, this will hurt readability
and has a higher risk of introducing bugs in testing
because conditions in assertions will be inverted
from what most people are used to
(checking the negative rather than the positive case).

This change introduces the ability to write these assertions
using a testing.T-compatible library like testify.

To do this, we introduce a new `*L` type analogous to `*T`
with an API similar to `testing.T`.
It implements more than sufficient methods to satisfy
Testify's TestingT interface, allowing its use with testify.

    // The following is all that is necessary for Testify.
    type L struct{ /* ... */ }
    func (*L) Errorf(string, ...any)
    func (*L) FailNow()

A function `WithL` is provided that constructs an `*L`,
runs the provided function, and returns the result.

    func WithL(func(*L)) LResult

`L` should not be instantiated directly, without this function,
as this function handles the goroutine management necessary
to make `FailNow()` halt execution for `require`-based assertions.

Finally, this alters the contract for these tests, slightly.
Previously, the result of a test was represented roughly by the type:

    type Result = enum {
        Success,
        Failure: { message: string }
    }

That is, either a test is successful, or it failed and has a message.

With this change, the result is represented with the type,

    type Result = {
        messages: []string,
        status: enum { Success, Failure }
    }

That is, messages may be included regardless of success or failure.
This brings it more in line with how `go test` works:
`t.Log` and `t.Error` both log messages,
one just marks the test as failed.

As of right now, messages are sent back to the test runner
at the end of the language test.
In a future change, we should explore streaming these messages back
into the `testing.T`.