Use the yaml converter plugin rather than Eject
We add some hooks to ensure we always use the version of
pulumi-converter-yaml that matches the version of yaml we've linked the
CLI to, this preserves the behaviour we have today that eject code
matches the gen code.

At some point it will make sense to decouple these and often just
default to the the latest converter instead.