Allow language plugins to return plugin checksums
Also fix an issue where if a platform was missing a checksum it would
error with "invalid checksum, expected , actual 01234".

None of the language runtimes yet return anything for this, but it's a
simple plumbing to expose it for the future.

We'll _probably_ start adding checksums to the pulumi-plugin.json files,
and then GetRequiredPlugins can simply return that data.